+++ api/package.json
@@ -9,7 +9,9 @@
     "@aws-sdk/client-dynamodb": "^3.110.0",
     "@aws-sdk/client-sqs": "^3.121.0",
     "@aws-sdk/lib-dynamodb": "^3.112.0",
+    "@multiformats/multiaddr": "^11.4.0",
     "aws-sdk": "^2.1145.0",
+    "is-local-ip": "^1.1.0",
     "multiaddr": "^10.0.1",
     "nanoid": "^4.0.0",
     "node-fetch": "^3.3.0",
