+++ pickup/package.json
@@ -17,6 +17,7 @@
     "@aws-sdk/client-s3": "^3.95.0",
     "@aws-sdk/lib-dynamodb": "^3.112.0",
     "@aws-sdk/lib-storage": "^3.97.0",
+    "@multiformats/multiaddr": "^11.4.0",
     "async-retry": "1.3.3",
     "debounce": "^1.2.1",
     "multiaddr": "^10.0.1",
